backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.6
py4j==0.10.9.7
pyparsing==3.0.9
rdflib==6.3.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
plotly==5.16.1
weasyprint==59.0
matplotlib==3.7.2
pandas==2.0.3
kaleido==0.2.1
python-dotenv==0.14.0
gunicorn==20.0.4
Werkzeug==2.3.6
